+++ packages/engine/src/component.rs
@@ -4,16 +4,15 @@ use utils::config::ChainConfigs;
 use wasmtime_wasi::{IoView, WasiCtx, WasiView};
 use wasmtime_wasi_http::{WasiHttpCtx, WasiHttpView};
 use wasmtime_wasi_keyvalue::WasiKeyValueCtx;
-use wavs_types::{Digest, ServiceID, Workflow, WorkflowID};
+use wavs_types::{Digest, Service, ServiceID, WorkflowID};
 
 use crate::bindings::world::host::LogLevel;
 
 // TODO: revisit this an understand it.
 // Copied blindly from old code
 pub struct HostComponent {
-    pub workflow: Workflow,
+    pub service: Service,
     pub workflow_id: WorkflowID,
-    pub service_id: ServiceID,
     pub chain_configs: ChainConfigs,
     pub(crate) table: wasmtime::component::ResourceTable,
     pub(crate) ctx: WasiCtx,