+++ packages/engine/src/error.rs
@@ -12,8 +12,14 @@ pub enum EngineError {
     #[error("Component returned an error: {0}")]
     ComponentError(anyhow::Error),
 
-    #[error{"Unable to get trigger data as component input: {0}"}]
-    TriggerData(anyhow::Error),
+    #[error("Workflow {workflow_id} not found for service {service_id}")]
+    WorkflowNotFound {
+        service_id: ServiceID,
+        workflow_id: WorkflowID,
+    },
+
+    #[error{"Unable to get component input: {0}"}]
+    Input(anyhow::Error),
 
     #[error{"Filesystem: {0}"}]
     Filesystem(anyhow::Error),